Add server integration, change font, and clean up JavaScript
Replace hard-coded data with server integration and update any functions that use or reference the hard-coded data to access the server instead.

Removed code that is no longer necessary since the data has moved to the server and re-organized the code to group related things together in an effort the make the code easier to navigate. Also created a couple of extra functions to help reduce the size (and improve the readability) of some of the anonymous functions. Helper functions that are used in at least two places or are not directly related to anything else in the script file are now grouped together at the bottom of the file.

Changed the font used to more closely match the Figma prototype provided by Gumroad.