Integrate the product ID
Added integration to use the product ID assigned to "The Minimalist Entrepreneur" in the database. This isn't strictly necessary based on the requirements of the challenge, but it is what I would do if I was making a more general ratings page, so I did it anyway. I also reworked the server API integration to match the changes I made to the server and added extra code to the load event to retrieve the product ID before retrieving the reviews.